raftstore: extract more coprocessor::properties
